ADD REFACTORED CODE
